Chatbot/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
beautifulsoup4
pypdf
chromadb
faiss-cpu
PyPDF2